+++ Changelog.md
@@ -1,3 +1,5 @@
 # 0.1.1
 - Add Optional Rate limiter
 - Update readme abit
+- Copy new readme to lib.rs
+- Add internal docs to Ratelimit